fgets() mock now rewinds after EOF
This project has code which reads through a GDB log file in two passes.
It reads it once to figure out what memory regions should be created
and then again to populate the simulated memory with values taken from
the log.  The modification I made to the fgets() mock has it return
NULL when it encounteres the last line in the test data and then
rewind to start again at the beginning of the test data.  This allows
a program under test to call fgets() until it gets NULL returned,
perform a simulated fseek() back to the beginning of the file and then
read it for a second time.