Add GdbLogParse() API
A user can use GDB commands to dump the contents of RAM and registers
when they are debugging a issue.  The results of these commands can be
logged and the resulting log is equivalent to a core dump that you
would see on other platforms.

The GdbLogParse() API can be used to read such a log file and use it
to initialize a IMemory object with the contents of RAM and a
RegisterContext object with the register contents.  These objects
can then be queried by GDB via the MRI core interface in CrashDebug to
allow post-mortem debugging at another location and/or time.