MemorySim_LoadFromFlashImage() now takes base address
Previously this function assumed that the FLASH binary image file
should always be loaded at a base address of FLASH_BASE_ADDRESS
(0x00000000).  The caller now has to specify this base address
explicitly.

I was testing CrashDebug with some old Smoothie dumps that I had
sitting around and discovered that it didn't work properly because
the .bin needed to be loaded at address 0x4000 since Smoothie uses
a bootloader in the lower portion of FLASH.