Grow GdbLogParser line buffer to 1024 chars
The sample Smoothie crash dumps that I had on my machine contained
lines longer than the old 128 width because of the symbols in the
memory lines.  My future recommended GDB commands to create such a
dump won't include these symbols so that the lines are a much smaller
fixed length.