Add ElfLoad module
This module is able to load read-only program entries from an ELF file
into read-only entries in the memory simulator.  With this module it is
possible to simulate the FLASH contents during a crash directly from
the .elf rather than requiring a .bin.

I added an elfFormatException to try_catch.h  This exception is thrown
if ElfLoad encounters anything in the .elf that it doesn't like.

I had to update the makefile as well to add libCrashDebug/src as an
include directory when building the tests so that it could access the
ElfPriv.h header file.  This header contains structure definitions for
the ELF standard.