CrashCatcherDump.c should use CrashCatcherMemoryRegionInfo
The CrashCatcher dumps only include the starting and ending address for
each memory region.  The elementSize isn't used.  Previously the code
in CrashCatcherDump.c which read in the dump files incorrectly assumed
that the dumps did include this elementSize field.  I needed to fix
this incorrect assumption in the unit tests as well.