Now able to load CrashCatcher dumps in CrashDebug
I updated the CrashDebugCommandLine module so that it can now load dump
data obtained from CrashCatcher in addition to the GDB logs that it
previously supported.  There are no new command line flags.  The --dump
option still takes the name of the dump file and the module looks at
the file's signature to see which loader/parser should be used.