Parse raw float register values in GDB logs
Floating point registers dumped in GDB look like:
"s1             1	(raw 0x3f800000)"

Previously the GDBLogParser module was parsing the first value
following the register name but this is the human readable floating
point value.  The parser should really use the raw hexadecimal value
that follows since it is the exact IEEE bit representation for the
single precision floating point value in the register.