Wait for GDB to connect
Now CrashDebug is configured to wait for GDB connection on first run.
This means that it won't try to send the T packet or fault description
text to GDB when GDB is first trying to send commands to learn more
about the debug stub and target. This reduces the time it takes for GDB
to startup since it no longer times out on the initial attempt it
makes to query the stub.
* Updated isGdbConnected() method in StandardIComm module to return
  FALSE until stdin has a byte ready to read or a byte has already been
  successfully read from stdin. At that point, we know that GDB is
  connected and communicating with this debug stub.
* I modified the mriPlatformBase test class to reconfigure
  Platform_CommShouldWaitForGdbConnect() to return FALSE like it did
  previously so that fewer tests needed to be modified. It also allows
  for T packet and fault description tests cases to still be run. Just
  a few tests set it to TRUE so that they can test how critical paths
  function in this default state.