Update memory read tests
Previously the MRI core would return an empty packet $#00 for the 'm'
command if the address supplied didn't point to valid memory. GDB
actually expects an error code to be returned in this situation. If a
read result is encountered after the first byte then it will still
return a truncated result like it did before. Just the case of 0 bytes
successfully read is special cased.