Add version 3.0 CrashCatcher support
The latest version of CrashCatcher, version 3.0, now includes the
contents of the MSP & PSP stack registers in its crash dumps.
CrashDebug now supports loading these new crash dumps and exposes the
MSP and PSP registers to GDB. The values for MSP & PSP returned to GDB
will either be the values found in version 3.0 or newer crash dumps or
a default value of 0xBAADBAAD when used with older 2.0 crash dumps.
This means that the newest CrashDebug, version 3, can load either
version 2.0 or version 3.0 CrashCatcher generated dumps.

Other updates in this commit include:
* Update the CrashCatcher submodule to the latest 3.0 version.
* Update the MRI submodule to the latest 1.0 version.
* Updated unit tests to verify the new functionality and maintain
  100% test code coverage.