Send all user output to stderr
Fixes issue #18.

Previously CrashDebug sent most of the output to stdout when it
encountered something on the command line which it didn't like. GDB
thinks that text sent to stdout is debug protocol chatter and gets
confused. It also doesn't get displayed to the user to let them know
that they have something wrong on the command line. Now the
CrashDebugCommandLine module sends all diagnostic output to stderr
instead which will be displayed in the GDB output so that the user
can determine what they need to correct.

Bumped version up to 1.7-20210805.0